test: share importable_tmp_path fixture, drop manual sys.path/modules cleanup
Promote the importable_tmp_path fixture from test_mkdocs_ext.py to
conftest.py so all test modules can use it, and refactor the #910 docs
tests to rely on it instead of hand-rolling sys.path.insert/remove +
sys.modules.pop inside try/finally.
