+++ tests/test_sphinx_ext.py
@@ -1051,7 +1051,7 @@ def serve(port: int = 8000):
             if "test_usage_name_module" in sys.modules:
                 del sys.modules["test_usage_name_module"]
 
-    def test_unnamed_app_uses_module_name_not_argv(self, tmp_path, monkeypatch):
+    def test_unnamed_app_uses_module_name_not_argv(self, importable_tmp_path, monkeypatch):
         """An unnamed command-group app derives its docs name from its module.
 
         ``App.name`` otherwise falls back to ``Path(sys.argv[0]).name``, which in
@@ -1060,7 +1060,7 @@ def serve(port: int = 8000):
         # Simulate a Sphinx build: sys.argv[0] is the generator, not the CLI.
         monkeypatch.setattr(sys, "argv", ["/usr/bin/sphinx-build", *sys.argv[1:]])
 
-        module_file = tmp_path / "myapptool.py"
+        module_file = importable_tmp_path / "myapptool.py"
         module_file.write_text(
             dedent(
                 """\
@@ -1076,44 +1076,37 @@ def serve(port: int = 8000):
             )
         )
 
-        sys.path.insert(0, str(tmp_path))
-        try:
-            from cyclopts.ext.sphinx import CycloptsDirective
+        from cyclopts.ext.sphinx import CycloptsDirective
 
-            mock_state = MagicMock()
-            mock_state.nested_parse = MagicMock()
+        mock_state = MagicMock()
+        mock_state.nested_parse = MagicMock()
 
-            directive = CycloptsDirective(
-                name="cyclopts",
-                arguments=["myapptool:app"],
-                options={},
-                content=StringList(),
-                lineno=1,
-                content_offset=0,
-                block_text="",
-                state=mock_state,
-                state_machine=MagicMock(),
-            )
-            directive.run()
+        directive = CycloptsDirective(
+            name="cyclopts",
+            arguments=["myapptool:app"],
+            options={},
+            content=StringList(),
+            lineno=1,
+            content_offset=0,
+            block_text="",
+            state=mock_state,
+            state_machine=MagicMock(),
+        )
+        directive.run()
 
-            rst_content = "\n".join(
-                line for call in mock_state.nested_parse.call_args_list if call for line in call[0][0]
-            )
-            assert "sphinx-build" not in rst_content
-            assert "myapptool serve" in rst_content  # Usage: line
-            assert ".. _cyclopts-myapptool-serve:" in rst_content  # anchor
-        finally:
-            sys.path.remove(str(tmp_path))
-            sys.modules.pop("myapptool", None)
+        rst_content = "\n".join(line for call in mock_state.nested_parse.call_args_list if call for line in call[0][0])
+        assert "sphinx-build" not in rst_content
+        assert "myapptool serve" in rst_content  # Usage: line
+        assert ".. _cyclopts-myapptool-serve:" in rst_content  # anchor
 
-    def test_no_duplicate_root_label_across_directives(self, tmp_path):
+    def test_no_duplicate_root_label_across_directives(self, importable_tmp_path):
         """Directives for different commands of one app must not share a root label.
 
         Every ``.. cyclopts::`` used to emit an identical ``cyclopts-<app>`` root
         label, so documenting different commands of the same app across pages
         produced Sphinx "duplicate label" warnings (#910).
         """
-        module_file = tmp_path / "dupmod.py"
+        module_file = importable_tmp_path / "dupmod.py"
         module_file.write_text(
             dedent(
                 """\
@@ -1134,39 +1127,34 @@ def serve(port: int = 8000):
             )
         )
 
-        sys.path.insert(0, str(tmp_path))
-        try:
-            from cyclopts.ext.sphinx import CycloptsDirective
+        from cyclopts.ext.sphinx import CycloptsDirective
 
-            def render(command: str) -> str:
-                mock_state = MagicMock()
-                mock_state.nested_parse = MagicMock()
-                directive = CycloptsDirective(
-                    name="cyclopts",
-                    arguments=["dupmod:app"],
-                    options={"commands": command},
-                    content=StringList(),
-                    lineno=1,
-                    content_offset=0,
-                    block_text="",
-                    state=mock_state,
-                    state_machine=MagicMock(),
-                )
-                directive.run()
-                return "\n".join(line for call in mock_state.nested_parse.call_args_list if call for line in call[0][0])
+        def render(command: str) -> str:
+            mock_state = MagicMock()
+            mock_state.nested_parse = MagicMock()
+            directive = CycloptsDirective(
+                name="cyclopts",
+                arguments=["dupmod:app"],
+                options={"commands": command},
+                content=StringList(),
+                lineno=1,
+                content_offset=0,
+                block_text="",
+                state=mock_state,
+                state_machine=MagicMock(),
+            )
+            directive.run()
+            return "\n".join(line for call in mock_state.nested_parse.call_args_list if call for line in call[0][0])
 
-            page_x = render("make-x")
-            page_y = render("make-y")
+        page_x = render("make-x")
+        page_y = render("make-y")
 
-            # No shared root label between the two pages (the duplicate-label source).
-            assert ".. _cyclopts-myapp:" not in page_x
-            assert ".. _cyclopts-myapp:" not in page_y
-            # Each page keeps its own unique, referenceable command anchor.
-            assert ".. _cyclopts-myapp-make-x:" in page_x
-            assert ".. _cyclopts-myapp-make-y:" in page_y
-        finally:
-            sys.path.remove(str(tmp_path))
-            sys.modules.pop("dupmod", None)
+        # No shared root label between the two pages (the duplicate-label source).
+        assert ".. _cyclopts-myapp:" not in page_x
+        assert ".. _cyclopts-myapp:" not in page_y
+        # Each page keeps its own unique, referenceable command anchor.
+        assert ".. _cyclopts-myapp-make-x:" in page_x
+        assert ".. _cyclopts-myapp-make-y:" in page_y
 
 
 class TestRstContentParsing:
